normalization addition and various spelling etc
